requirements.txt
bottle==0.11.6
docutils==0.11
pytz==2013.8
IPython>=1.0
jinja2==2.7.1
pyyaml==3.10
